requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
async-timeout==5.0.1
attrs==21.2.0
Automat==20.2.0
bcrypt==3.2.0
beautifulsoup4==4.13.3
blinker==1.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
chardet==4.0.0
charset-normalizer==3.4.1
click==8.0.3
cloud-init==19.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
Deprecated==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.2
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-genai==1.7.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==21.3.0
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.1.3
Markdown==3.7
MarkupSafe==2.0.1
more-itertools==8.10.0
multidict==6.1.0
ncatbot==3.8.10.post5
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.0
pexpect==4.8.0
pillow==11.1.0
playwright==1.50.0
propcache==0.3.0
psutil==6.1.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pydantic==2.10.6
pydantic_core==2.27.2
pyee==11.1.1
Pygments==2.19.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyppeteer==2.0.0
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2025.2
PyYAML==6.0.2
qrcode==8.0
qrcode-terminal==0.8
requests==2.32.3
rsa==4.9
schedule==1.2.2
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sos==4.5.6
soupsieve==2.6
sseclient-py==1.8.0
ssh-import-id==5.11
style==1.1.0
systemd-python==234
tqdm==4.67.1
Twisted==22.1.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
update==0.0.1
urllib3==1.26.20
wadllib==1.3.6
websockets==10.4
wrapt==1.17.2
xkit==0.0.0
yarl==1.18.3
zipp==3.21.0
zope.interface==5.4.0
